Revert introduction of Respond.js
Respond.js while it seemed was a good idea, was in fact very troublesome,
because our CSS is hosted on CDN. Since don't have control of servers, it made
for a very difficult set-up and poor results.

We are reverting to revious technique. That is hardcoding a non-responsive
stylesheet for IE8 only.