Prevent scroll jump when showing last items
In Chrome, not Firefox, the whole page would scroll way too down when
showing the last items of the "more" button. We can workaround this by
using somehting else than hide(), or queue hide. But we also want to keep
the button layout so that the content below never jump. This is even better
than before and works on both Firefox and Chrome